+++ changelog.md
@@ -6,6 +6,7 @@
 * Bugfix uncaught AccessDeniedException for patches on top-level interface atoms
 * Improve stack trace. Now also showing trace of previous errors/exceptions
 * Add config for setting data dir (needed for containerizing Ampersand backend application). Uploads folder is now fixed relative to data folder
+* Add option to set certain configurations via environment variables. Starting with AMPERSAND_DEBUG_MODE and AMPERSAND_DBHOST
 
 ## v1.4.0 (3 january 2020)
 * Bugfix issue with API interfaces shown in UI to solve signal violations. Caused by wrongly placed parentheses.