+++ changelog.md
@@ -7,6 +7,7 @@
 * Improve stack trace. Now also showing trace of previous errors/exceptions
 * Add config for setting data dir (needed for containerizing Ampersand backend application). Uploads folder is now fixed relative to data folder
 * Add option to set certain configurations via environment variables. Starting with AMPERSAND_DEBUG_MODE and AMPERSAND_DBHOST
+* Simplify ways to specify configurations. Removed recent option of environment folder structure (was introduced in v1.4.0)
 
 ## v1.4.0 (3 january 2020)
 * Bugfix issue with API interfaces shown in UI to solve signal violations. Caused by wrongly placed parentheses.