+++ changelog.md
@@ -10,6 +10,7 @@
 * Simplify ways to specify configurations. Removed recent option of environment folder structure (was introduced in v1.4.0)
 * OAuthLogin extension: allow to specify urls in config file relative to global.serverURL
 * Bugfix. When loading configuration files, first load extensions and after that additional config files
+* Add devcontainer configuration to repository
 
 ## v1.4.0 (3 january 2020)
 * Bugfix issue with API interfaces shown in UI to solve signal violations. Caused by wrongly placed parentheses.